# openapi/firemail.yaml
...
openapi: 3.0.3
info:
  title: FireMailPlus API
  version: 0.3.0
  license:
    name: Proprietary
  description: |
    Contract-first API definition for the stable FireMailPlus `/api/v1`
    backend surface.

    This Phase 1 contract intentionally includes only routes that are
    registered in the production Gin router. Dead, unregistered, fake-success,
    or schema-blocked handlers remain excluded until their repair tasks are
    completed.
servers:
  - url: /
    description: Relative server URL
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/emails/search:
    get:
      tags: [Emails]
      operationId: searchEmails
      summary: Search emails
      parameters:
        - name: q
          in: query
          required: false
          schema:
            type: string
        - name: subject
          in: query
          required: false
          schema:
            type: string
        - name: from
          in: query
          required: false
          schema:
            type: string
        - name: to
          in: query
          required: false
          schema:
            type: string
        - name: body
          in: query
          required: false
          schema:
            type: string
        - name: since
          in: query
          required: false
          schema:
            type: string
            format: date-time
        - name: before
          in: query
          required: false
          schema:
            type: string
            format: date-time
        - name: has_attachment
          in: query
          required: false
          schema:
            type: boolean
        - name: is_read
          in: query
          required: false
          schema:
            type: boolean
        - name: is_starred
          in: query
          required: false
          schema:
            type: boolean
        - $ref: '#/components/parameters/AccountIdQuery'
        - $ref: '#/components/parameters/FolderIdQuery'
        - $ref: '#/components/parameters/PageQuery'
        - $ref: '#/components/parameters/PageSizeQuery'
      responses:
        '200':
          $ref: '#/components/responses/EmailsPageSuccess'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/emails/{id}:
    get:
      tags: [Emails]
      operationId: getEmail
      summary: Get email detail
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          $ref: '#/components/responses/EmailSuccess'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Emails]
      operationId: updateEmail
      summary: Update email flags
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        $ref: '#/components/requestBodies/JsonBody'
      responses:
        '200':
          $ref: '#/components/responses/EmailSuccess'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Emails]
      operationId: deleteEmail
      summary: Delete email
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          $ref: '#/components/responses/Success'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/emails/send:
    post:
      tags: [Emails]
      operationId: sendEmail
      summary: Send email
      requestBody:
        $ref: '#/components/requestBodies/JsonBody'
      responses:
        '200':
          $ref: '#/components/responses/Success'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/emails/{id}/read:
    put:
      tags: [Emails]
      operationId: markEmailAsRead
      summary: Mark email as read
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          $ref: '#/components/responses/Success'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
        '502':
          $ref: '#/components/responses/BadGateway'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/emails/{id}/unread:
    put:
      tags: [Emails]
      operationId: markEmailAsUnread
      summary: Mark email as unread
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          $ref: '#/components/responses/Success'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
        '502':
          $ref: '#/components/responses/BadGateway'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/emails/{id}/star:
    put:
      tags: [Emails]
      operationId: toggleEmailStar
      summary: Toggle email star
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          $ref: '#/components/responses/EmailSuccess'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/emails/{id}/move:
    put:
      tags: [Emails]
      operationId: moveEmail
      summary: Move email
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        $ref: '#/components/requestBodies/JsonBody'
      responses:
        '200':
          $ref: '#/components/responses/Success'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/emails/{id}/archive:
    put:
      tags: [Emails]
      operationId: archiveEmail
      summary: Archive email
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          $ref: '#/components/responses/Success'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/emails/{id}/reply:
    post:
      tags: [Emails]
      operationId: replyEmail
      summary: Reply to email
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        $ref: '#/components/requestBodies/JsonBody'
      responses:
        '200':
          $ref: '#/components/responses/Success'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/emails/{id}/reply-all:
    post:
      tags: [Emails]
      operationId: replyAllEmail
      summary: Reply all to email
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        $ref: '#/components/requestBodies/JsonBody'
      responses:
        '200':
          $ref: '#/components/responses/Success'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/emails/{id}/forward:
    post:
      tags: [Emails]
      operationId: forwardEmail
      summary: Forward email
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        $ref: '#/components/requestBodies/JsonBody'
      responses:
        '200':
          $ref: '#/components/responses/Success'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/emails/batch:
    post:
      tags: [Emails]
      operationId: batchEmailOperations
      summary: Batch email operations
      requestBody:
        $ref: '#/components/requestBodies/JsonBody'
      responses:
        '200':
          $ref: '#/components/responses/Success'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    queryToken:
      type: apiKey
      in: query
      name: token
      description: Browser EventSource compatibility token. Server logs must redact this value.
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
...
    AccountIdQuery:
      name: account_id
      in: query
      required: false
      schema:
        type: integer
        format: int64
        minimum: 1
    FolderIdQuery:
      name: folder_id
      in: query
      required: false
      schema:
        type: integer
        format: int64
        minimum: 1
    PageQuery:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSizeQuery:
      name: page_size
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 20
...
  requestBodies:
    JsonBody:
      required: true
      content:
        application/json:
          schema:
            type: object
            additionalProperties: true
  responses:
    Success:
      description: Operation succeeded.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/SuccessResponse'
...
    EmailsPageSuccess:
      description: Email page response.
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/SuccessResponse'
              - type: object
                properties:
                  data:
                    $ref: '#/components/schemas/EmailPage'
    EmailSuccess:
      description: Email response.
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/SuccessResponse'
              - type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Email'
...
    BadRequest:
      description: Invalid request.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Missing or invalid authentication.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    Conflict:
      description: Request conflicts with the current resource state.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    BadGateway:
      description: Upstream mailbox provider operation failed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    ErrorResponse:
      type: object
      additionalProperties: true
      required: [error]
      properties:
        error:
          type: string
        message:
          type: string
        code:
          type: string
    SuccessResponse:
      type: object
      additionalProperties: false
      required: [success]
      properties:
        success:
          type: boolean
        data:
          type: object
          nullable: true
          additionalProperties: true
        message:
          type: string
...
    EmailAddress:
      type: object
      additionalProperties: true
      properties:
        name:
          type: string
        address:
          type: string
          format: email
...
    Email:
      type: object
      additionalProperties: true
      properties:
        id:
          type: integer
          format: int64
        account_id:
          type: integer
          format: int64
        folder_id:
          type: integer
          format: int64
        subject:
          type: string
        from:
          $ref: '#/components/schemas/EmailAddress'
        to:
          type: array
          items:
            $ref: '#/components/schemas/EmailAddress'
        cc:
          type: array
          items:
            $ref: '#/components/schemas/EmailAddress'
        bcc:
          type: array
          items:
            $ref: '#/components/schemas/EmailAddress'
        text_body:
          type: string
        html_body:
          type: string
        is_read:
          type: boolean
        is_starred:
          type: boolean
        is_important:
          type: boolean
        has_attachment:
          type: boolean
        date:
          type: string
    EmailPage:
      type: object
      additionalProperties: true
      properties:
        emails:
          type: array
          items:
            $ref: '#/components/schemas/Email'
        total:
          type: integer
        page:
          type: integer
        page_size:
          type: integer
        total_pages:
          type: integer
...